+++ libraries/muxer/src/test/java/androidx/media3/muxer/BoxesTest.java
@@ -517,8 +517,7 @@ public class BoxesTest {
   @Test
   public void
       convertPresentationTimestampsToDurationsVu_singleSampleAtZeroTimestamp_returnsSampleLengthEqualsZero() {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
-        createBufferInfoListWithSamplePresentationTimestamps(0L);
+    List<BufferInfo> sampleBufferInfos = createBufferInfoListWithSamplePresentationTimestamps(0L);
 
     List<Integer> durationsVu =
         Boxes.convertPresentationTimestampsToDurationsVu(
@@ -533,7 +532,7 @@ public class BoxesTest {
   @Test
   public void
       convertPresentationTimestampsToDurationsVu_singleSampleAtNonZeroTimestamp_returnsSampleLengthEqualsZero() {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
+    List<BufferInfo> sampleBufferInfos =
         createBufferInfoListWithSamplePresentationTimestamps(5_000L);
 
     List<Integer> durationsVu =
@@ -549,7 +548,7 @@ public class BoxesTest {
   @Test
   public void
       convertPresentationTimestampsToDurationsVu_differentSampleDurations_lastFrameDurationShort_returnsLastSampleOfZeroDuration() {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
+    List<BufferInfo> sampleBufferInfos =
         createBufferInfoListWithSamplePresentationTimestamps(0L, 30_000L, 80_000L);
 
     List<Integer> durationsVu =
@@ -565,7 +564,7 @@ public class BoxesTest {
   @Test
   public void
       convertPresentationTimestampsToDurationsVu_differentSampleDurations_lastFrameDurationDuplicate_returnsLastSampleOfDuplicateDuration() {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
+    List<BufferInfo> sampleBufferInfos =
         createBufferInfoListWithSamplePresentationTimestamps(0L, 30_000L, 80_000L);
 
     List<Integer> durationsVu =
@@ -581,7 +580,7 @@ public class BoxesTest {
   @Test
   public void
       convertPresentationTimestampsToDurationsVu_withOutOfOrderSampleTimestamps_returnsExpectedDurations() {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
+    List<BufferInfo> sampleBufferInfos =
         createBufferInfoListWithSamplePresentationTimestamps(0L, 10_000L, 1_000L, 2_000L, 11_000L);
 
     List<Integer> durationsVu =
@@ -597,7 +596,7 @@ public class BoxesTest {
   @Test
   public void
       convertPresentationTimestampsToDurationsVu_withLastSampleDurationBehaviorUsingEndOfStreamFlag_returnsExpectedDurations() {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
+    List<BufferInfo> sampleBufferInfos =
         createBufferInfoListWithSamplePresentationTimestamps(0L, 1_000L, 2_000L, 3_000L, 4_000L);
 
     List<Integer> durationsVu =
@@ -652,8 +651,7 @@ public class BoxesTest {
 
   @Test
   public void createCttsBox_withSingleSampleTimestamp_returnsEmptyBox() {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
-        createBufferInfoListWithSamplePresentationTimestamps(400);
+    List<BufferInfo> sampleBufferInfos = createBufferInfoListWithSamplePresentationTimestamps(400);
     List<Integer> durationsVu =
         Boxes.convertPresentationTimestampsToDurationsVu(
             sampleBufferInfos,
@@ -669,7 +667,7 @@ public class BoxesTest {
 
   @Test
   public void createCttsBox_withNoBframesSampleTimestamps_returnsEmptyBox() throws IOException {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
+    List<BufferInfo> sampleBufferInfos =
         createBufferInfoListWithSamplePresentationTimestamps(0L, 1000L, 2000L);
     List<Integer> durationsVu =
         Boxes.convertPresentationTimestampsToDurationsVu(
@@ -686,7 +684,7 @@ public class BoxesTest {
 
   @Test
   public void createCttsBox_withBFramesSampleTimestamps_matchesExpected() throws IOException {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
+    List<BufferInfo> sampleBufferInfos =
         createBufferInfoListWithSamplePresentationTimestamps(
             0, 400, 200, 100, 300, 800, 600, 500, 700);
 
@@ -706,7 +704,7 @@ public class BoxesTest {
 
   @Test
   public void createCttsBox_withLargeSampleTimestamps_matchesExpected() throws IOException {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
+    List<BufferInfo> sampleBufferInfos =
         createBufferInfoListWithSamplePresentationTimestamps(
             23698215060L, 23698248252L, 23698347988L, 23698488968L, 23698547416L);
 
@@ -724,8 +722,7 @@ public class BoxesTest {
 
   @Test
   public void createStszBox_matchesExpected() throws IOException {
-    List<MediaCodec.BufferInfo> sampleBufferInfos =
-        createBufferInfoListWithSampleSizes(100, 200, 150, 200);
+    List<BufferInfo> sampleBufferInfos = createBufferInfoListWithSampleSizes(100, 200, 150, 200);
 
     ByteBuffer stszBox = Boxes.stsz(sampleBufferInfos);
 
@@ -780,7 +777,7 @@ public class BoxesTest {
 
   @Test
   public void createStssBox_matchesExpected() throws IOException {
-    List<MediaCodec.BufferInfo> sampleBufferInfos = createBufferInfoListWithSomeKeyFrames();
+    List<BufferInfo> sampleBufferInfos = createBufferInfoListWithSomeKeyFrames();
 
     ByteBuffer stssBox = Boxes.stss(sampleBufferInfos);
 
@@ -867,37 +864,35 @@ public class BoxesTest {
         context, dumpableBox, MuxerTestUtil.getExpectedDumpFilePath("trex_box"));
   }
 
-  private static List<MediaCodec.BufferInfo> createBufferInfoListWithSamplePresentationTimestamps(
+  private static List<BufferInfo> createBufferInfoListWithSamplePresentationTimestamps(
       long... timestampsUs) {
-    List<MediaCodec.BufferInfo> bufferInfoList = new ArrayList<>();
+    List<BufferInfo> bufferInfoList = new ArrayList<>();
     for (long timestampUs : timestampsUs) {
-      MediaCodec.BufferInfo bufferInfo = new MediaCodec.BufferInfo();
-      bufferInfo.presentationTimeUs = timestampUs;
+      BufferInfo bufferInfo = new BufferInfo(timestampUs, /* size= */ 0, /* flags= */ 0);
       bufferInfoList.add(bufferInfo);
     }
 
     return bufferInfoList;
   }
 
-  private static List<MediaCodec.BufferInfo> createBufferInfoListWithSampleSizes(int... sizes) {
-    List<MediaCodec.BufferInfo> bufferInfoList = new ArrayList<>();
+  private static List<BufferInfo> createBufferInfoListWithSampleSizes(int... sizes) {
+    List<BufferInfo> bufferInfoList = new ArrayList<>();
     for (int size : sizes) {
-      MediaCodec.BufferInfo bufferInfo = new MediaCodec.BufferInfo();
-      bufferInfo.size = size;
+      BufferInfo bufferInfo = new BufferInfo(/* presentationTimeUs= */ 0, size, /* flags= */ 0);
       bufferInfoList.add(bufferInfo);
     }
 
     return bufferInfoList;
   }
 
-  private static List<MediaCodec.BufferInfo> createBufferInfoListWithSomeKeyFrames() {
-    List<MediaCodec.BufferInfo> bufferInfoList = new ArrayList<>();
+  private static List<BufferInfo> createBufferInfoListWithSomeKeyFrames() {
+    List<BufferInfo> bufferInfoList = new ArrayList<>();
     for (int i = 0; i < 30; i++) {
-      MediaCodec.BufferInfo bufferInfo = new MediaCodec.BufferInfo();
+      @C.BufferFlags int flags = 0;
       if (i % 5 == 0) { // Make every 5th frame as key frame.
-        bufferInfo.flags = MediaCodec.BUFFER_FLAG_KEY_FRAME;
+        flags = C.BUFFER_FLAG_KEY_FRAME;
       }
-      bufferInfoList.add(bufferInfo);
+      bufferInfoList.add(new BufferInfo(/* presentationTimeUs= */ 0, /* size= */ 0, flags));
     }
 
     return bufferInfoList;